Updated the Script to handle multiple branches as per OP-20456

+++ upgrade/isd-generate-yamls.sh
@@ -61,7 +61,7 @@ sed -i 's/^data:/stringData:/' /repo/oes/templates/sapor-gate/sapor-gate-secret.
 sed -i 's/{{ .Values.saporgate.config.password }}/encrypted:saporpassword:saporpassword/' /repo/oes/config/sapor-gate/gate-local.yml
 ####################################################################################################################
 helm template ${release} /repo/oes/ -f values.yaml --output-dir=/tmp/isd
-if [ $? -eq 0 ]; then  
+if [ $? -eq 0 ]; then
      echo "#################################Helm template is sucessfull into isd directory#################################"
 else
    echo "#################################Helm template failed to isd directory#################################"
@@ -74,25 +74,55 @@ rm -rf /tmp/isd/oes/templates/hooks/cleanup.yaml
 rm -rf /repo/oes/
 rm -rf oes-"$version".tgz
 #####################################committing tempates to github repo################################
-git branch "$version"
-if [ $? -eq 0 ]; then  
-     echo "#################################Sucesfully created a branch in github#################################"
-else
-   echo "#################################Failed to created a branch in github#################################"
-   exit 1
+git branch -a > branch
+checkbranch=$(cat branch | grep "$version" | awk -F "/" '{print $3}')
+if [ -z "$checkbranch" ]; then
+   echo "#################################Sucesfully created a branch in github###########################"
+   git checkout -b "$version"
+   git rm -rf /repo/isd/
+   cp -r /tmp/isd/ /repo/
+   rm -rf /repo/branch
+   git status
+   git add .
+   git config --global user.email "${gitemail}"
+   git config --global user.name "${username}"
+   git commit -m "Manifest file dir of helm chart with ISD ${version}"
+   git push origin "$version"
+   if [ $? -eq 0 ]; then
+     echo "#######################Sucessfully pushed helm template to github#################################"
+   else
+     echo "#########################Failed to pushed helm template to github#################################"
+     exit 1
+   fi
+ else
+   echo "#################################create a branch in github with hipens###############################"
+   cat branch | grep "$version"| awk -F "/" '{print $3}' | grep "-"
+   if [ $? == 0 ]; then
+      echo "######### Creating a new branch with hipens because already exists with $version #####################"
+      i=$(cat branch | grep "$version"| awk -F "/" '{print $3}' | grep -v HEAD | grep "-" | awk -F "-" '{print $2}'| tail -1)
+      i=$((i+1))
+      newbranch="$version-$i"
+      echo $newbranch
+      echo "Incremented the branch with hipens $newbranch"
+   else
+     echo "############ Creating a new branch with $version-1 since $version branch exists##############"
+     newbranch="$version-1"
+   fi
+   git checkout -b "$newbranch"
+   git rm -rf /repo/isd/
+   rm -rf /repo/branch
+   cp -r /tmp/isd/ /repo/
+   git status
+   git add .
+   git config --global user.email "${gitemail}"
+   git config --global user.name "${username}"
+   git commit -m "Manifest file dir of helm chart with $newbranch since branch with ISD ${version} exists"
+   git push origin "$newbranch"
+   if [ $? -eq 0 ]; then
+     echo "#######################Sucessfully pushed helm template to github#################################"
+   else
+     echo "#########################Failed to pushed helm template to github#################################"
+     exit 1
+   fi
 fi
-git checkout "$version"
-git rm -rf /repo/isd/
-cp -r /tmp/isd/ /repo/
-git status
-git add .
-git config --global user.email "${gitemail}"
-git config --global user.name "${username}"
-git commit -m "Manifest file dir of helm chart with ISD ${version}"
-git push origin "$version" --force
-if [ $? -eq 0 ]; then  
-     echo "#################################Sucesfully pushed helm template to github#################################"
-else
-   echo "#################################Failed to pushed helm template to github#################################"
-   exit 1
-fi
+#############################################################
